Add enhanced mode for CloseWithoutCloseableRule. Introduce base class for dual mode rules called AbstractEnhanceableAstVisitorRule. Allow enabling enhanced mode for all dual mode rules using a system property. Simplify testing of dual mode rules by introducing GenericAbstractRuleTestCase.
Supersedes PR #27.
