+++ src/main/groovy/org/codenarc/rule/design/CompareToWithoutComparableRule.groovy
@@ -16,8 +16,7 @@
 package org.codenarc.rule.design
 
 import org.codehaus.groovy.ast.MethodNode
-import org.codehaus.groovy.control.Phases
-import org.codenarc.rule.AbstractAstVisitorRule
+import org.codenarc.rule.AbstractEnhanceableAstVisitorRule
 import org.codenarc.rule.AbstractMethodVisitor
 import org.codenarc.util.AstUtil
 
@@ -28,16 +27,10 @@ import org.codenarc.util.AstUtil
  *
  * @author Hamlet D'Arcy
  */
-class CompareToWithoutComparableRule extends AbstractAstVisitorRule {
+class CompareToWithoutComparableRule extends AbstractEnhanceableAstVisitorRule {
     String name = 'CompareToWithoutComparable'
     int priority = 2
     Class astVisitorClass = CompareToWithoutComparableAstVisitor
-    boolean enhancedMode
-
-    @Override
-    int getCompilerPhase() {
-        enhancedMode ? Phases.CANONICALIZATION : super.compilerPhase
-    }
 }
 
 class CompareToWithoutComparableAstVisitor extends AbstractMethodVisitor {
